Complete part 1 of assignment
The added Containerfile defines a deployable OCI-compliant image for the
application.

In several places, the build or image could have been cleaner with minor
edits to the build script or app. For example, `build.sh` hardcodes the
`virtualenv` command, which is a third-party dependency. For simple
Python projects, it would be cleaner to replace it with `python -m
venv`, but other build depenencies aren't so simple and I left it in for
the purpose of this assignment. The build script also contains no error
handling, which isn't ideal but fixing it is outside the scope of this
assignment.

The image was built and tested using Podman, which I prefer over Docker
because the rootless implementation is cleaner in my experience.