Complete part 2 of assignment
Create a trivial kubernetes deployment using minikube and kustomize.

I used minikube out of personal familiarity. It's not a platform that
I would ever use in production, so I have no strong opinions on it other
than that it works on my computer.

Kustomize, on the other hand, I chose after evaluating several
alternatives, each of which I have some experience with in more serious
settings.

The most straightforward option would have been to just use bare KRM
with no frills. For an application like this it would have sufficed, but
I chose to approach this assignment in a way that reflects how I would
real-world problems and editing dozens of KRM files with no abstraction
is not a good workflow.

Helm is one of the most popular solutions, but also one I personally
dislike quite strongly. At any modicum of scale, it becomes hostile to
automation. Coming from years of experience with Ansible and Jinja2,
I can confidently say that a full-featured templating engine is the
wrong choice of technology for configuration files. The ecosystem also
has a similar quality control problem as Ansible Galaxy or NPM, severely
limiting the one advantage it has over other tools.

I have used kpt on larger projects and enjoyed it greatly. It follows
a similar paradigm as kustomize with more fleshed out support for
programmatic changes and GitOps. On a more complex project I would
probably have used kpt, but for this assignment I opted for the simpler
tool that does not add a third-party dependency.