Move cloudbuild configs to repository
GCP has an extremely convenient way to set up a cloudbuild-cloudrun
pipeline in only a few clicks on the GUI, but it's not available in
Terraform yet: https://cloud.google.com/run/docs/quickstarts/deploy-continuously
This means that easiest way to set this up is to initialize it in the
GUI and import each part into version control.

I'm moving the cloudbuild configs first so that when I redeploy
all the cloud infrastructure it should immediately work instead of
having to parse out known errors from user mistakes.