Complete part 3 of assignment
I implemented automated build/test/deployment using a pipeline of GCP
services: Cloud Build, Artifact Registry, and Cloud Run. The application
container is automatically built and runs unittests whenever a PR is
opened against a matching branch (dev/master in this case). The results
are sent back to GitHub for easy review. Once the patch is merged,
a separate workflow kicks off. It rebuilds the image, deploys it to
a private repository, then deploys it to Cloud Run. Like with the unit
tests, the status of this process is communicated to GitHub so the user
can see the change go live as soon as the green checkmark appears.

I chose this pipeline for a few reasons. It was communicated to me that
Formlabs is intending to focus on GCP infrastructure, and this is
a demonstration of what GCP-based CI might look like. For Cloud Run in
particular, I've found it to be the easiest production-ready
kubernetes-based platform by far (it's knative under the hood). Despite
its relative immaturity, I've had very good experience using it for
deploying standard web apps quickly. Cloud Build and Artifact Registry,
on the other hand, are bog-standard. If you're using GCP, you're using
those.

There are tradeoffs for this CI solution. Most glaringly, the container
image is rebuilt for deployment when a merge occurs. This could be
problematic if the build process is time-consuming or
compute-intensive. One way to optimize this would be to iron out strict
Github branch permissions to ensure nothing gets merged outside of the
PR process, and tag the built images with the PR number. The push
trigger would then be able to correlate the merge commit and the
matching label and re-use that. Of course, this kind of added complexity
has its own downsides that should be considered on balance.

The Terraform infrastructure for this app is a little condensed. For
more complex systems, I would usually separate out the dev, prod, and
other environments more explicitly. This is both to prevent deployment
errors and becaues different environments often have deeper differences
than a few tags. I kept them together for this app so that it would be
easier to review. The more robust structure also has less marginal
utility without its own CI system to support it (I like using Atlantis
for that).